Pass options to Discover
This enables passing options to harmonyhubjs-discover's Discover
constructor.

This closes #105.

For systems with multiple interfaces, using the default broadcast
address of 255.255.255.255 will not give consistent results. Instead,
you need to pass the appropriate broadcast address for the network
you're addressing, like 192.168.1.255 for the network 192.168.1.0/24.

This PR requires swissmanu/harmonyhubjs-discover#8 to be merged and
released.